% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{block_alert}
\alias{block_alert}
\title{Create an alert}
\usage{
block_alert(
  title,
  ...,
  description = NULL,
  icon = "info",
  variant = c("default", "destructive"),
  class = NULL
)
}
\arguments{
\item{title}{Alert title. Required for accessibility.}

\item{...}{Additional alert body content.}

\item{description}{Optional alert description.}

\item{icon}{Optional icon tag or vendored icon name.}

\item{variant}{Visual variant.}

\item{class}{Additional classes.}
}
\value{
An \code{htmltools} tag.
}
\description{
Create an alert
}
\seealso{
Other content: 
\code{\link{block_alert_description}()},
\code{\link{block_alert_title}()},
\code{\link{block_badge}()},
\code{\link{block_card}()}
}
\concept{content}
