% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{block_alert_description}
\alias{block_alert_description}
\title{Create an alert description}
\usage{
block_alert_description(..., class = NULL)
}
\arguments{
\item{...}{Alert description content.}

\item{class}{Additional classes.}
}
\value{
An \code{htmltools} tag.
}
\description{
Create an alert description
}
\seealso{
Other content: 
\code{\link{block_alert}()},
\code{\link{block_alert_title}()},
\code{\link{block_badge}()},
\code{\link{block_card}()}
}
\concept{content}
