% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{block_alert_title}
\alias{block_alert_title}
\title{Create an alert title}
\usage{
block_alert_title(..., class = NULL)
}
\arguments{
\item{...}{Alert title content.}

\item{class}{Additional classes.}
}
\value{
An \code{htmltools} tag.
}
\description{
Create an alert title
}
\seealso{
Other content: 
\code{\link{block_alert}()},
\code{\link{block_alert_description}()},
\code{\link{block_badge}()},
\code{\link{block_card}()}
}
\concept{content}
