% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{block_badge}
\alias{block_badge}
\title{Create a badge}
\usage{
block_badge(
  label,
  variant = c("default", "secondary", "outline", "destructive"),
  class = NULL
)
}
\arguments{
\item{label}{Badge label.}

\item{variant}{Visual variant.}

\item{class}{Additional classes.}
}
\value{
An \code{htmltools} tag.
}
\description{
Create a badge
}
\seealso{
Other content: 
\code{\link{block_alert}()},
\code{\link{block_alert_description}()},
\code{\link{block_alert_title}()},
\code{\link{block_card}()},
\code{\link{block_card_content}()},
\code{\link{block_card_description}()},
\code{\link{block_card_footer}()},
\code{\link{block_card_header}()},
\code{\link{block_card_title}()},
\code{\link{block_dialog}()},
\code{\link{block_empty}()},
\code{\link{block_popover}()},
\code{\link{block_separator}()},
\code{\link{block_skeleton}()},
\code{\link{block_spinner}()},
\code{\link{block_tooltip}()},
\code{\link{block_value_box}()},
\code{\link{update_block_dialog}()},
\code{\link{update_block_popover}()}
}
\concept{content}
