% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/card.R
\name{block_card_header}
\alias{block_card_header}
\title{Create a card header}
\usage{
block_card_header(..., class = NULL)
}
\arguments{
\item{...}{Card header content.}

\item{class}{Additional classes.}
}
\value{
An \code{htmltools} tag.
}
\description{
Create a card header
}
\seealso{
Other content: 
\code{\link{block_alert}()},
\code{\link{block_alert_action}()},
\code{\link{block_alert_description}()},
\code{\link{block_alert_title}()},
\code{\link{block_badge}()},
\code{\link{block_card}()},
\code{\link{block_card_content}()},
\code{\link{block_card_description}()},
\code{\link{block_card_footer}()},
\code{\link{block_card_title}()},
\code{\link{block_code}()},
\code{\link{block_dialog}()},
\code{\link{block_empty}()},
\code{\link{block_popover}()},
\code{\link{block_separator}()},
\code{\link{block_skeleton}()},
\code{\link{block_spinner}()},
\code{\link{block_table}()},
\code{\link{block_toaster}()},
\code{\link{block_tooltip}()},
\code{\link{block_value_box}()},
\code{\link{dismiss_toast}()},
\code{\link{show_toast}()},
\code{\link{table_column}()},
\code{\link{update_block_dialog}()},
\code{\link{update_block_popover}()},
\code{\link{update_block_table}()},
\code{\link{update_block_toaster}()}
}
\concept{content}
