% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dark-mode.R
\name{block_dark_mode_toggle}
\alias{block_dark_mode_toggle}
\title{Create a dark mode toggle}
\usage{
block_dark_mode_toggle(label = "Theme", class = NULL)
}
\arguments{
\item{label}{Button label.}

\item{class}{Additional classes.}
}
\value{
An \code{htmltools} tag.
}
\description{
Create a dark mode toggle
}
\seealso{
Other theme:
\code{\link[=block_style]{block_style()}},
\code{\link[=block_style_profiles]{block_style_profiles()}},
\code{\link[=block_theme]{block_theme()}},
\code{\link[=block_theme_presets]{block_theme_presets()}},
\code{\link[=update_block_theme]{update_block_theme()}}
}
\concept{theme}
