% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicators.R
\name{block_empty}
\alias{block_empty}
\title{Create an empty state}
\usage{
block_empty(
  title,
  ...,
  description = NULL,
  icon = NULL,
  action = NULL,
  class = NULL,
  style = NULL
)
}
\arguments{
\item{title}{Empty-state title.}

\item{...}{Additional empty-state body content.}

\item{description}{Optional description.}

\item{icon}{Optional icon tag or vendored icon name.}

\item{action}{Optional action content.}

\item{class}{Additional classes.}

\item{style}{Optional inline custom styles.}
}
\value{
An \code{htmltools} tag.
}
\description{
Create an empty state
}
\seealso{
Other content: 
\code{\link{block_alert}()},
\code{\link{block_alert_action}()},
\code{\link{block_alert_description}()},
\code{\link{block_alert_title}()},
\code{\link{block_badge}()},
\code{\link{block_card}()},
\code{\link{block_card_content}()},
\code{\link{block_card_description}()},
\code{\link{block_card_footer}()},
\code{\link{block_card_header}()},
\code{\link{block_card_title}()},
\code{\link{block_code}()},
\code{\link{block_dialog}()},
\code{\link{block_popover}()},
\code{\link{block_separator}()},
\code{\link{block_skeleton}()},
\code{\link{block_spinner}()},
\code{\link{block_table}()},
\code{\link{block_tooltip}()},
\code{\link{block_value_box}()},
\code{\link{table_column}()},
\code{\link{update_block_dialog}()},
\code{\link{update_block_popover}()}
}
\concept{content}
