% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/field.R
\name{block_field_description}
\alias{block_field_description}
\title{Create field helper text}
\usage{
block_field_description(..., id = NULL, class = NULL)
}
\arguments{
\item{...}{Description content.}

\item{id}{Optional element id.}

\item{class}{Additional classes.}
}
\value{
An \code{htmltools} tag.
}
\description{
Create field helper text
}
\seealso{
Other forms: 
\code{\link{block_checkbox}()},
\code{\link{block_field}()},
\code{\link{block_field_group}()},
\code{\link{block_field_invalid}()},
\code{\link{block_field_label}()},
\code{\link{block_field_legend}()},
\code{\link{block_field_set}()},
\code{\link{block_input}()},
\code{\link{block_input_group}()},
\code{\link{block_input_group_addon}()},
\code{\link{block_radio_group}()},
\code{\link{block_select}()},
\code{\link{block_slider}()},
\code{\link{block_switch}()},
\code{\link{block_textarea}()},
\code{\link{update_block_checkbox}()},
\code{\link{update_block_input}()},
\code{\link{update_block_radio_group}()},
\code{\link{update_block_select}()},
\code{\link{update_block_slider}()},
\code{\link{update_block_switch}()},
\code{\link{update_block_textarea}()}
}
\concept{forms}
