% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/field.R
\name{block_field_group}
\alias{block_field_group}
\title{Create a field group}
\usage{
block_field_group(..., class = NULL)
}
\arguments{
\item{...}{Field content.}

\item{class}{Additional classes.}
}
\value{
An \code{htmltools} tag.
}
\description{
Create a field group
}
\seealso{
Other forms: 
\code{\link{block_checkbox}()},
\code{\link{block_field}()},
\code{\link{block_field_description}()},
\code{\link{block_field_invalid}()},
\code{\link{block_field_label}()},
\code{\link{block_field_legend}()},
\code{\link{block_field_set}()},
\code{\link{block_input_group}()},
\code{\link{block_input_group_addon}()},
\code{\link{block_select}()},
\code{\link{block_slider}()},
\code{\link{block_switch}()},
\code{\link{block_textarea}()}
}
\concept{forms}
