% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/field.R
\name{block_field_label}
\alias{block_field_label}
\title{Create a field label}
\usage{
block_field_label(..., `for` = NULL, class = NULL)
}
\arguments{
\item{...}{Label content.}

\item{for}{Optional input id referenced by the label.}

\item{class}{Additional classes.}
}
\value{
An \code{htmltools} tag.
}
\description{
Create a field label
}
\seealso{
Other forms: 
\code{\link{block_checkbox}()},
\code{\link{block_field}()},
\code{\link{block_field_description}()},
\code{\link{block_field_group}()},
\code{\link{block_field_invalid}()},
\code{\link{block_field_legend}()},
\code{\link{block_field_set}()},
\code{\link{block_input_group}()},
\code{\link{block_input_group_addon}()},
\code{\link{block_select}()},
\code{\link{block_slider}()},
\code{\link{block_switch}()},
\code{\link{block_textarea}()},
\code{\link{update_block_select}()}
}
\concept{forms}
