% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-group.R
\name{block_input_group_addon}
\alias{block_input_group_addon}
\title{Create an input group addon}
\usage{
block_input_group_addon(..., class = NULL)
}
\arguments{
\item{...}{Addon content.}

\item{class}{Additional classes.}
}
\value{
An \code{htmltools} tag.
}
\description{
Create an input group addon
}
\seealso{
Other forms: 
\code{\link{block_checkbox}()},
\code{\link{block_field}()},
\code{\link{block_field_description}()},
\code{\link{block_field_group}()},
\code{\link{block_field_invalid}()},
\code{\link{block_field_label}()},
\code{\link{block_field_legend}()},
\code{\link{block_field_set}()},
\code{\link{block_input_group}()},
\code{\link{block_select}()},
\code{\link{block_slider}()},
\code{\link{block_switch}()},
\code{\link{block_textarea}()},
\code{\link{update_block_checkbox}()},
\code{\link{update_block_select}()},
\code{\link{update_block_switch}()},
\code{\link{update_block_textarea}()}
}
\concept{forms}
