% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{block_sidebar}
\alias{block_sidebar}
\title{Create a dashboard sidebar}
\usage{
block_sidebar(..., title = NULL)
}
\arguments{
\item{...}{Sidebar content.}

\item{title}{Optional sidebar title.}
}
\value{
An \code{htmltools} tag.
}
\description{
Create a dashboard sidebar
}
