% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicators.R
\name{block_skeleton}
\alias{block_skeleton}
\title{Create a skeleton placeholder}
\usage{
block_skeleton(class = NULL, ...)
}
\arguments{
\item{class}{Additional classes.}

\item{...}{Additional attributes passed to \code{htmltools::tags$div}.}
}
\value{
An \code{htmltools} tag.
}
\description{
Create a skeleton placeholder
}
\seealso{
Other content: 
\code{\link{block_alert}()},
\code{\link{block_alert_action}()},
\code{\link{block_alert_description}()},
\code{\link{block_alert_title}()},
\code{\link{block_badge}()},
\code{\link{block_card}()},
\code{\link{block_card_content}()},
\code{\link{block_card_description}()},
\code{\link{block_card_footer}()},
\code{\link{block_card_header}()},
\code{\link{block_card_title}()},
\code{\link{block_code}()},
\code{\link{block_dialog}()},
\code{\link{block_empty}()},
\code{\link{block_popover}()},
\code{\link{block_separator}()},
\code{\link{block_spinner}()},
\code{\link{block_tooltip}()},
\code{\link{block_value_box}()},
\code{\link{update_block_dialog}()},
\code{\link{update_block_popover}()}
}
\concept{content}
