% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{block_spinner}
\alias{block_spinner}
\title{Create a spinner}
\usage{
block_spinner(label = "Loading", class = NULL)
}
\arguments{
\item{label}{Accessible label.}

\item{class}{Additional classes.}
}
\value{
An \code{htmltools} tag.
}
\description{
Create a spinner
}
\seealso{
Other content: 
\code{\link{block_alert}()},
\code{\link{block_alert_description}()},
\code{\link{block_alert_title}()},
\code{\link{block_badge}()},
\code{\link{block_card}()},
\code{\link{block_card_content}()},
\code{\link{block_card_description}()},
\code{\link{block_card_footer}()},
\code{\link{block_card_header}()},
\code{\link{block_card_title}()},
\code{\link{block_dialog}()},
\code{\link{block_empty}()},
\code{\link{block_separator}()},
\code{\link{block_skeleton}()},
\code{\link{block_value_box}()}
}
\concept{content}
