% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style.R
\name{block_style_profiles}
\alias{block_style_profiles}
\title{Supported built-in style profiles}
\usage{
block_style_profiles()
}
\value{
A character vector of built-in style-profile names accepted by
\code{\link[=block_style]{block_style()}}.
}
\description{
Supported built-in style profiles
}
\examples{
block_style_profiles()
}
\seealso{
Other theme:
\code{\link[=block_dark_mode_toggle]{block_dark_mode_toggle()}},
\code{\link[=block_style]{block_style()}},
\code{\link[=block_theme]{block_theme()}},
\code{\link[=block_theme_presets]{block_theme_presets()}},
\code{\link[=update_block_theme]{update_block_theme()}}
}
\concept{theme}
