% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabs.R
\name{block_tab}
\alias{block_tab}
\title{Create a tab}
\usage{
block_tab(title, ..., value = NULL)
}
\arguments{
\item{title}{Tab label.}

\item{...}{Tab content.}

\item{value}{Optional tab value.}
}
\value{
An \code{htmltools} tag.
}
\description{
Create a tab
}
\seealso{
Other navigation:
\code{\link[=block_nav]{block_nav()}},
\code{\link[=block_nav_group]{block_nav_group()}},
\code{\link[=block_nav_item]{block_nav_item()}},
\code{\link[=block_nav_label]{block_nav_label()}},
\code{\link[=block_tabs]{block_tabs()}},
\code{\link[=update_block_nav]{update_block_nav()}},
\code{\link[=update_block_tabs]{update_block_tabs()}}
}
\concept{navigation}
