% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{block_theme}
\alias{block_theme}
\title{Create page-scoped theme overrides}
\usage{
block_theme(...)
}
\arguments{
\item{...}{Named CSS token overrides, such as \code{primary},
\code{background}, or \code{radius}.}
}
\value{
An \code{htmltools} tag.
}
\description{
Create page-scoped theme overrides
}
\seealso{
Other theme: 
\code{\link{block_dark_mode_toggle}()},
\code{\link{update_block_theme}()}
}
\concept{theme}
