% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{block_theme_presets}
\alias{block_theme_presets}
\title{Supported built-in colour presets}
\usage{
block_theme_presets()
}
\value{
A character vector of built-in palette names accepted by the
\code{preset} argument of \code{\link[=block_theme]{block_theme()}}.
}
\description{
Supported built-in colour presets
}
\examples{
block_theme_presets()
}
\seealso{
Other theme:
\code{\link[=block_dark_mode_toggle]{block_dark_mode_toggle()}},
\code{\link[=block_style]{block_style()}},
\code{\link[=block_style_profiles]{block_style_profiles()}},
\code{\link[=block_theme]{block_theme()}},
\code{\link[=update_block_theme]{update_block_theme()}}
}
\concept{theme}
