% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{block_value_box}
\alias{block_value_box}
\title{Create a value box}
\usage{
block_value_box(
  title,
  value,
  ...,
  description = NULL,
  icon = NULL,
  class = NULL
)
}
\arguments{
\item{title}{Value box title.}

\item{value}{Primary value.}

\item{...}{Additional value box body content.}

\item{description}{Optional value box description.}

\item{icon}{Optional icon tag or vendored icon name.}

\item{class}{Additional classes.}
}
\value{
An \code{htmltools} tag.
}
\description{
Create a value box
}
\seealso{
Other content: 
\code{\link{block_alert}()},
\code{\link{block_alert_description}()},
\code{\link{block_alert_title}()},
\code{\link{block_badge}()},
\code{\link{block_card}()},
\code{\link{block_card_content}()},
\code{\link{block_card_description}()},
\code{\link{block_card_footer}()},
\code{\link{block_card_header}()},
\code{\link{block_card_title}()},
\code{\link{block_empty}()},
\code{\link{block_separator}()},
\code{\link{block_skeleton}()},
\code{\link{block_spinner}()}
}
\concept{content}
