% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{shadcn_header}
\alias{shadcn_header}
\title{Create a dashboard header}
\usage{
shadcn_header(...)
}
\arguments{
\item{...}{Header content.}
}
\value{
An \code{htmltools} tag.
}
\description{
Create a dashboard header
}
