% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{shinyshadcn-package}
\alias{shinyshadcn}
\alias{shinyshadcn-package}
\title{shinyshadcn: Shadcn-inspired dashboards for Shiny}
\description{
\code{shinyshadcn} provides composable dashboard primitives for Shiny apps.
}
\author{
\strong{Maintainer}: Niels van der Velden \email{n.s.j.vandervelden@gmail.com}

}
\keyword{internal}
