% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{update_block_select}
\alias{update_block_select}
\title{Update a runtime select input}
\usage{
update_block_select(
  session = shiny::getDefaultReactiveDomain(),
  input_id,
  selected,
  choices,
  placeholder,
  disabled,
  notify = FALSE
)
}
\arguments{
\item{session}{Shiny session. Defaults to the current reactive domain.}

\item{input_id}{Input id passed to \code{block_select()}.}

\item{selected}{Optional selected value. \code{NULL} clears the value.}

\item{choices}{Optional replacement choices.}

\item{placeholder}{Optional replacement placeholder.}

\item{disabled}{Optional disabled state.}

\item{notify}{Whether Shiny should receive an input event for \code{selected}.}
}
\value{
Invisibly returns \code{NULL}.
}
\description{
Update a runtime select input
}
\seealso{
Other forms: 
\code{\link{block_checkbox}()},
\code{\link{block_field}()},
\code{\link{block_field_description}()},
\code{\link{block_field_group}()},
\code{\link{block_field_invalid}()},
\code{\link{block_field_label}()},
\code{\link{block_field_legend}()},
\code{\link{block_field_set}()},
\code{\link{block_input_group}()},
\code{\link{block_input_group_addon}()},
\code{\link{block_select}()},
\code{\link{block_slider}()},
\code{\link{block_switch}()},
\code{\link{block_textarea}()}
}
\concept{forms}
