% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{update_block_theme}
\alias{update_block_theme}
\title{Update the active theme}
\usage{
update_block_theme(
  session = shiny::getDefaultReactiveDomain(),
  mode = c("system", "light", "dark")
)
}
\arguments{
\item{session}{Shiny session.}

\item{mode}{Theme mode: \code{system}, \code{light}, or \code{dark}.}
}
\value{
Invisibly returns \code{NULL}.
}
\description{
Update the active theme
}
\seealso{
Other theme: 
\code{\link{block_dark_mode_toggle}()},
\code{\link{block_theme}()}
}
\concept{theme}
