% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{block_alert}
\alias{block_alert}
\title{Create an alert}
\usage{
block_alert(
  title,
  ...,
  description = NULL,
  action = NULL,
  icon = "info",
  variant = c("default", "destructive", "success", "warning", "info"),
  class = NULL,
  style = NULL
)
}
\arguments{
\item{title}{Alert title. Required for accessibility.}

\item{...}{Additional alert body content.}

\item{description}{Optional alert description.}

\item{action}{Optional action content, such as a \code{block_button()}.}

\item{icon}{Optional icon tag or vendored icon name.}

\item{variant}{Visual variant.}

\item{class}{Additional classes.}

\item{style}{Optional inline custom styles.}
}
\value{
An \code{htmltools} tag.
}
\description{
Create an alert
}
\seealso{
Other content: 
\code{\link{block_alert_action}()},
\code{\link{block_alert_description}()},
\code{\link{block_alert_title}()},
\code{\link{block_badge}()},
\code{\link{block_card}()},
\code{\link{block_card_content}()},
\code{\link{block_card_description}()},
\code{\link{block_card_footer}()},
\code{\link{block_card_header}()},
\code{\link{block_card_title}()},
\code{\link{block_code}()},
\code{\link{block_dialog}()},
\code{\link{block_empty}()},
\code{\link{block_popover}()},
\code{\link{block_separator}()},
\code{\link{block_skeleton}()},
\code{\link{block_spinner}()},
\code{\link{block_tooltip}()},
\code{\link{block_value_box}()},
\code{\link{update_block_dialog}()},
\code{\link{update_block_popover}()}
}
\concept{content}
