% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alert.R
\name{block_alert_title}
\alias{block_alert_title}
\title{Create an alert title}
\usage{
block_alert_title(..., class = NULL)
}
\arguments{
\item{...}{Alert title content.}

\item{class}{Additional classes.}
}
\value{
An \code{htmltools} tag.
}
\description{
Create an alert title
}
\seealso{
Other content: 
\code{\link{block_alert}()},
\code{\link{block_alert_action}()},
\code{\link{block_alert_description}()},
\code{\link{block_badge}()},
\code{\link{block_card}()},
\code{\link{block_card_content}()},
\code{\link{block_card_description}()},
\code{\link{block_card_footer}()},
\code{\link{block_card_header}()},
\code{\link{block_card_title}()},
\code{\link{block_code}()},
\code{\link{block_dialog}()},
\code{\link{block_empty}()},
\code{\link{block_popover}()},
\code{\link{block_separator}()},
\code{\link{block_skeleton}()},
\code{\link{block_spinner}()},
\code{\link{block_table}()},
\code{\link{block_tooltip}()},
\code{\link{block_value_box}()},
\code{\link{table_column}()},
\code{\link{update_block_dialog}()},
\code{\link{update_block_popover}()},
\code{\link{update_block_table}()}
}
\concept{content}
