% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{block_badge}
\alias{block_badge}
\title{Create a badge}
\usage{
block_badge(
  label,
  variant = c("default", "secondary", "outline", "destructive"),
  class = NULL
)
}
\arguments{
\item{label}{Badge label.}

\item{variant}{Visual variant.}

\item{class}{Additional classes.}
}
\value{
An \code{htmltools} tag.
}
\description{
Create a badge
}
\seealso{
Other content: 
\code{\link{block_alert}()},
\code{\link{block_alert_description}()},
\code{\link{block_alert_title}()},
\code{\link{block_card}()}
}
\concept{content}
