% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/page.R
\name{block_body}
\alias{block_body}
\title{Create a page body landmark}
\usage{
block_body(..., class = NULL)
}
\arguments{
\item{...}{Body content.}

\item{class}{Additional classes.}
}
\value{
An \code{htmltools} tag.
}
\description{
Create a page body landmark
}
\seealso{
Other layout: 
\code{\link{block_cluster}()},
\code{\link{block_grid}()},
\code{\link{block_header}()},
\code{\link{block_page}()},
\code{\link{block_sidebar}()},
\code{\link{block_stack}()}
}
\concept{layout}
