% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{block_card}
\alias{block_card}
\title{Create a dashboard card}
\usage{
block_card(..., title = NULL, value = NULL, class = NULL)
}
\arguments{
\item{...}{Card body content.}

\item{title}{Optional card title.}

\item{value}{Optional primary value.}

\item{class}{Additional classes.}
}
\value{
An \code{htmltools} tag.
}
\description{
Create a dashboard card
}
\concept{content}
