% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{block_card}
\alias{block_card}
\title{Create a dashboard card}
\usage{
block_card(
  ...,
  title = NULL,
  description = NULL,
  value = NULL,
  footer = NULL,
  class = NULL
)
}
\arguments{
\item{...}{Card body content or composed card region tags.}

\item{title}{Optional card title.}

\item{description}{Optional card description.}

\item{value}{Optional primary value.}

\item{footer}{Optional card footer content.}

\item{class}{Additional classes.}
}
\value{
An \code{htmltools} tag.
}
\description{
Create a dashboard card
}
\seealso{
Other content: 
\code{\link{block_alert}()},
\code{\link{block_alert_description}()},
\code{\link{block_alert_title}()},
\code{\link{block_badge}()},
\code{\link{block_card_content}()},
\code{\link{block_card_description}()},
\code{\link{block_card_footer}()},
\code{\link{block_card_header}()},
\code{\link{block_card_title}()},
\code{\link{block_code}()},
\code{\link{block_dialog}()},
\code{\link{block_empty}()},
\code{\link{block_popover}()},
\code{\link{block_separator}()},
\code{\link{block_skeleton}()},
\code{\link{block_spinner}()},
\code{\link{block_tooltip}()},
\code{\link{block_value_box}()},
\code{\link{update_block_dialog}()},
\code{\link{update_block_popover}()}
}
\concept{content}
