% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dark-mode.R
\name{block_dark_mode_toggle}
\alias{block_dark_mode_toggle}
\title{Create a dark mode toggle}
\usage{
block_dark_mode_toggle(label = "Theme", class = NULL)
}
\arguments{
\item{label}{Button label.}

\item{class}{Additional classes.}
}
\value{
An \code{htmltools} tag.
}
\description{
Create a dark mode toggle
}
\seealso{
Other theme: 
\code{\link{block_theme}()},
\code{\link{update_block_theme}()}
}
\concept{theme}
