% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/field.R
\name{block_field_invalid}
\alias{block_field_invalid}
\title{Mark a field invalid}
\usage{
block_field_invalid(field, message)
}
\arguments{
\item{field}{A \code{block_field()} tag.}

\item{message}{Validation message shown below the control.}
}
\value{
A modified \code{htmltools} tag.
}
\description{
Mark a field invalid
}
\seealso{
Other forms: 
\code{\link{block_checkbox}()},
\code{\link{block_date_picker}()},
\code{\link{block_field}()},
\code{\link{block_field_description}()},
\code{\link{block_field_group}()},
\code{\link{block_field_label}()},
\code{\link{block_field_legend}()},
\code{\link{block_field_set}()},
\code{\link{block_file_input}()},
\code{\link{block_input}()},
\code{\link{block_input_group}()},
\code{\link{block_input_group_addon}()},
\code{\link{block_radio_group}()},
\code{\link{block_select}()},
\code{\link{block_slider}()},
\code{\link{block_switch}()},
\code{\link{block_textarea}()},
\code{\link{update_block_checkbox}()},
\code{\link{update_block_date_picker}()},
\code{\link{update_block_file_input}()},
\code{\link{update_block_input}()},
\code{\link{update_block_radio_group}()},
\code{\link{update_block_select}()},
\code{\link{update_block_slider}()},
\code{\link{update_block_switch}()},
\code{\link{update_block_textarea}()}
}
\concept{forms}
