% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/field.R
\name{block_field_legend}
\alias{block_field_legend}
\title{Create a field set legend}
\usage{
block_field_legend(..., class = NULL)
}
\arguments{
\item{...}{Legend content.}

\item{class}{Additional classes.}
}
\value{
An \code{htmltools} tag.
}
\description{
Create a field set legend
}
\seealso{
Other forms:
\code{\link[=block_checkbox]{block_checkbox()}},
\code{\link[=block_combobox]{block_combobox()}},
\code{\link[=block_date_picker]{block_date_picker()}},
\code{\link[=block_date_range_picker]{block_date_range_picker()}},
\code{\link[=block_field]{block_field()}},
\code{\link[=block_field_description]{block_field_description()}},
\code{\link[=block_field_group]{block_field_group()}},
\code{\link[=block_field_invalid]{block_field_invalid()}},
\code{\link[=block_field_label]{block_field_label()}},
\code{\link[=block_field_set]{block_field_set()}},
\code{\link[=block_file_input]{block_file_input()}},
\code{\link[=block_input]{block_input()}},
\code{\link[=block_input_group]{block_input_group()}},
\code{\link[=block_input_group_addon]{block_input_group_addon()}},
\code{\link[=block_radio_group]{block_radio_group()}},
\code{\link[=block_select]{block_select()}},
\code{\link[=block_slider]{block_slider()}},
\code{\link[=block_switch]{block_switch()}},
\code{\link[=block_textarea]{block_textarea()}},
\code{\link[=block_toggle_group]{block_toggle_group()}},
\code{\link[=update_block_checkbox]{update_block_checkbox()}},
\code{\link[=update_block_combobox]{update_block_combobox()}},
\code{\link[=update_block_date_picker]{update_block_date_picker()}},
\code{\link[=update_block_date_range_picker]{update_block_date_range_picker()}},
\code{\link[=update_block_file_input]{update_block_file_input()}},
\code{\link[=update_block_input]{update_block_input()}},
\code{\link[=update_block_radio_group]{update_block_radio_group()}},
\code{\link[=update_block_select]{update_block_select()}},
\code{\link[=update_block_slider]{update_block_slider()}},
\code{\link[=update_block_switch]{update_block_switch()}},
\code{\link[=update_block_textarea]{update_block_textarea()}},
\code{\link[=update_block_toggle_group]{update_block_toggle_group()}}
}
\concept{forms}
