% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{block_header}
\alias{block_header}
\title{Create a dashboard header}
\usage{
block_header(..., class = NULL)
}
\arguments{
\item{...}{Header content.}

\item{class}{Additional classes.}
}
\value{
An \code{htmltools} tag.
}
\description{
Create a dashboard header
}
\seealso{
Other layout:
\code{\link[=block_body]{block_body()}},
\code{\link[=block_cluster]{block_cluster()}},
\code{\link[=block_grid]{block_grid()}},
\code{\link[=block_page]{block_page()}},
\code{\link[=block_sidebar]{block_sidebar()}},
\code{\link[=block_stack]{block_stack()}}
}
\concept{layout}
