% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{block_nav}
\alias{block_nav}
\title{Create a navigation container}
\usage{
block_nav(..., class = NULL)
}
\arguments{
\item{...}{Navigation items.}

\item{class}{Additional classes.}
}
\value{
An \code{htmltools} tag.
}
\description{
Create a navigation container
}
\seealso{
Other navigation: 
\code{\link{block_nav_item}()}
}
\concept{navigation}
