% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/page.R
\name{block_page}
\alias{block_page}
\title{Create a modern Shiny page}
\usage{
block_page(..., title = NULL, sidebar = NULL, header = NULL)
}
\arguments{
\item{...}{Page body content.}

\item{title}{Browser page title.}

\item{sidebar}{Optional sidebar content.}

\item{header}{Optional header content.}
}
\value{
An \code{htmltools} tag list suitable for a Shiny UI.
}
\description{
Create a modern Shiny page
}
