% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{block_select}
\alias{block_select}
\title{Create a styled select input}
\usage{
block_select(
  input_id,
  choices,
  selected = NULL,
  placeholder = NULL,
  disabled = FALSE,
  width = NULL,
  class = NULL
)
}
\arguments{
\item{input_id}{Input id.}

\item{choices}{Choice labels and values.}

\item{selected}{Optional selected value.}

\item{placeholder}{Optional placeholder shown when no value is selected.}

\item{disabled}{Whether the control is disabled.}

\item{width}{Optional CSS width value.}

\item{class}{Additional classes.}
}
\value{
An \code{htmltools} tag.
}
\description{
Create a styled select input
}
\seealso{
Other forms: 
\code{\link{block_field}()},
\code{\link{block_field_description}()},
\code{\link{block_field_group}()},
\code{\link{block_field_invalid}()},
\code{\link{block_field_label}()},
\code{\link{block_field_legend}()},
\code{\link{block_field_set}()},
\code{\link{block_input_group}()},
\code{\link{block_input_group_addon}()}
}
\concept{forms}
