% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{block_separator}
\alias{block_separator}
\title{Create a separator}
\usage{
block_separator(
  orientation = c("horizontal", "vertical"),
  decorative = TRUE,
  class = NULL
)
}
\arguments{
\item{orientation}{Separator orientation.}

\item{decorative}{Whether the separator is decorative only.}

\item{class}{Additional classes.}
}
\value{
An \code{htmltools} tag.
}
\description{
Create a separator
}
\seealso{
Other content: 
\code{\link{block_alert}()},
\code{\link{block_alert_action}()},
\code{\link{block_alert_description}()},
\code{\link{block_alert_title}()},
\code{\link{block_badge}()},
\code{\link{block_card}()},
\code{\link{block_card_content}()},
\code{\link{block_card_description}()},
\code{\link{block_card_footer}()},
\code{\link{block_card_header}()},
\code{\link{block_card_title}()},
\code{\link{block_code}()},
\code{\link{block_dialog}()},
\code{\link{block_empty}()},
\code{\link{block_popover}()},
\code{\link{block_skeleton}()},
\code{\link{block_spinner}()},
\code{\link{block_tooltip}()},
\code{\link{block_value_box}()},
\code{\link{update_block_dialog}()},
\code{\link{update_block_popover}()}
}
\concept{content}
