% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{block_sidebar}
\alias{block_sidebar}
\title{Create a dashboard sidebar}
\usage{
block_sidebar(..., title = NULL, class = NULL)
}
\arguments{
\item{...}{Sidebar content.}

\item{title}{Optional sidebar title.}

\item{class}{Additional classes.}
}
\value{
An \code{htmltools} tag.
}
\description{
Create a dashboard sidebar
}
\seealso{
Other layout: 
\code{\link{block_body}()},
\code{\link{block_header}()},
\code{\link{block_page}()}
}
\concept{layout}
