% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{block_sidebar}
\alias{block_sidebar}
\title{Create a dashboard sidebar}
\usage{
block_sidebar(
  ...,
  title = NULL,
  collapsible = FALSE,
  collapsed = FALSE,
  id = NULL,
  class = NULL
)
}
\arguments{
\item{...}{Sidebar content.}

\item{title}{Optional sidebar title.}

\item{collapsible}{Whether the sidebar can collapse on larger screens.}

\item{collapsed}{Whether the sidebar starts collapsed on larger screens.}

\item{id}{Optional sidebar DOM id.}

\item{class}{Additional classes.}
}
\value{
An \code{htmltools} tag.
}
\description{
Create a dashboard sidebar
}
\seealso{
Other layout: 
\code{\link{block_body}()},
\code{\link{block_cluster}()},
\code{\link{block_grid}()},
\code{\link{block_header}()},
\code{\link{block_page}()},
\code{\link{block_stack}()}
}
\concept{layout}
