% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicators.R
\name{block_spinner}
\alias{block_spinner}
\title{Create a spinner}
\usage{
block_spinner(
  label = "Loading",
  size = c("default", "sm", "lg"),
  color = c("default", "muted", "primary", "destructive", "success", "warning", "info"),
  class = NULL,
  style = NULL
)
}
\arguments{
\item{label}{Accessible \code{aria-label} announced by assistive technology.}

\item{size}{Visual size.}

\item{color}{Semantic color.}

\item{class}{Additional classes.}

\item{style}{Optional inline custom styles.}
}
\value{
An \code{htmltools} tag.
}
\description{
Create a spinner
}
\seealso{
Other content: 
\code{\link{block_alert}()},
\code{\link{block_alert_action}()},
\code{\link{block_alert_description}()},
\code{\link{block_alert_title}()},
\code{\link{block_badge}()},
\code{\link{block_card}()},
\code{\link{block_card_content}()},
\code{\link{block_card_description}()},
\code{\link{block_card_footer}()},
\code{\link{block_card_header}()},
\code{\link{block_card_title}()},
\code{\link{block_code}()},
\code{\link{block_dialog}()},
\code{\link{block_empty}()},
\code{\link{block_popover}()},
\code{\link{block_separator}()},
\code{\link{block_skeleton}()},
\code{\link{block_table}()},
\code{\link{block_tooltip}()},
\code{\link{block_value_box}()},
\code{\link{table_column}()},
\code{\link{update_block_dialog}()},
\code{\link{update_block_popover}()}
}
\concept{content}
