% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/form-controls.R
\name{block_switch}
\alias{block_switch}
\title{Create a styled switch input}
\usage{
block_switch(
  input_id,
  label,
  value = FALSE,
  disabled = FALSE,
  style = NULL,
  class = NULL
)
}
\arguments{
\item{input_id}{Input id.}

\item{label}{Switch label.}

\item{value}{Whether the switch starts on.}

\item{disabled}{Whether the control is disabled.}

\item{style}{Inline CSS styles.}

\item{class}{Additional classes.}
}
\value{
An \code{htmltools} tag.
}
\description{
Create a styled switch input
}
\seealso{
Other forms: 
\code{\link{block_checkbox}()},
\code{\link{block_field}()},
\code{\link{block_field_description}()},
\code{\link{block_field_group}()},
\code{\link{block_field_invalid}()},
\code{\link{block_field_label}()},
\code{\link{block_field_legend}()},
\code{\link{block_field_set}()},
\code{\link{block_input_group}()},
\code{\link{block_input_group_addon}()},
\code{\link{block_select}()},
\code{\link{block_slider}()},
\code{\link{block_textarea}()},
\code{\link{update_block_checkbox}()},
\code{\link{update_block_select}()},
\code{\link{update_block_switch}()},
\code{\link{update_block_textarea}()}
}
\concept{forms}
