% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabs.R
\name{block_tabs}
\alias{block_tabs}
\title{Create styled tabs}
\usage{
block_tabs(
  ...,
  id = NULL,
  selected = NULL,
  variant = c("default", "line"),
  orientation = c("horizontal", "vertical"),
  class = NULL
)
}
\arguments{
\item{...}{\code{block_tab()} or \code{shiny::tabPanel()} items.}

\item{id}{Optional input id.}

\item{selected}{Optional selected tab value.}

\item{variant}{Visual variant: \code{default} or \code{line}.}

\item{orientation}{Layout orientation: \code{horizontal} or \code{vertical}.}

\item{class}{Additional classes.}
}
\value{
An \code{htmltools} tag.
}
\description{
Create styled tabs
}
\seealso{
Other navigation: 
\code{\link{block_nav}()},
\code{\link{block_nav_item}()},
\code{\link{block_tab}()},
\code{\link{update_block_tabs}()}
}
\concept{navigation}
