% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/form-controls.R
\name{block_textarea}
\alias{block_textarea}
\title{Create a styled textarea input}
\usage{
block_textarea(
  input_id,
  value = "",
  placeholder = NULL,
  rows = 2,
  width = NULL,
  disabled = FALSE,
  class = NULL
)
}
\arguments{
\item{input_id}{Input id.}

\item{value}{Initial value.}

\item{placeholder}{Optional placeholder text.}

\item{rows}{Number of visible rows.}

\item{width}{Optional CSS width value.}

\item{disabled}{Whether the control is disabled.}

\item{class}{Additional classes.}
}
\value{
An \code{htmltools} tag.
}
\description{
Create a styled textarea input
}
\seealso{
Other forms: 
\code{\link{block_checkbox}()},
\code{\link{block_field}()},
\code{\link{block_field_description}()},
\code{\link{block_field_group}()},
\code{\link{block_field_invalid}()},
\code{\link{block_field_label}()},
\code{\link{block_field_legend}()},
\code{\link{block_field_set}()},
\code{\link{block_input_group}()},
\code{\link{block_input_group_addon}()},
\code{\link{block_select}()},
\code{\link{block_slider}()},
\code{\link{block_switch}()},
\code{\link{update_block_select}()}
}
\concept{forms}
