% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toast.R
\name{dismiss_toast}
\alias{dismiss_toast}
\title{Dismiss a toast notification}
\usage{
dismiss_toast(
  session = shiny::getDefaultReactiveDomain(),
  toaster_id,
  toast_id = NULL
)
}
\arguments{
\item{session}{Shiny session. Defaults to the current reactive session.}

\item{toaster_id}{Target \code{\link[=block_toaster]{block_toaster()}} id.}

\item{toast_id}{Non-empty id of the toast to dismiss (as returned by
\code{\link[=show_toast]{show_toast()}}). \code{NULL} dismisses all toasts.}
}
\value{
Invisibly returns \code{NULL}.
}
\description{
Removes a toast from a \code{\link[=block_toaster]{block_toaster()}} before it auto-dismisses. With no
\code{toast_id}, clears every visible toast.
}
\seealso{
\code{\link[=block_toaster]{block_toaster()}}, \code{\link[=show_toast]{show_toast()}}

Other content:
\code{\link[=block_accordion]{block_accordion()}},
\code{\link[=block_accordion_item]{block_accordion_item()}},
\code{\link[=block_alert]{block_alert()}},
\code{\link[=block_alert_action]{block_alert_action()}},
\code{\link[=block_alert_description]{block_alert_description()}},
\code{\link[=block_alert_dialog]{block_alert_dialog()}},
\code{\link[=block_alert_title]{block_alert_title()}},
\code{\link[=block_badge]{block_badge()}},
\code{\link[=block_card]{block_card()}},
\code{\link[=block_card_content]{block_card_content()}},
\code{\link[=block_card_description]{block_card_description()}},
\code{\link[=block_card_footer]{block_card_footer()}},
\code{\link[=block_card_header]{block_card_header()}},
\code{\link[=block_card_title]{block_card_title()}},
\code{\link[=block_code]{block_code()}},
\code{\link[=block_dialog]{block_dialog()}},
\code{\link[=block_dropdown_menu]{block_dropdown_menu()}},
\code{\link[=block_empty]{block_empty()}},
\code{\link[=block_popover]{block_popover()}},
\code{\link[=block_progress]{block_progress()}},
\code{\link[=block_separator]{block_separator()}},
\code{\link[=block_skeleton]{block_skeleton()}},
\code{\link[=block_spinner]{block_spinner()}},
\code{\link[=block_table]{block_table()}},
\code{\link[=block_toaster]{block_toaster()}},
\code{\link[=block_tooltip]{block_tooltip()}},
\code{\link[=block_value_box]{block_value_box()}},
\code{\link[=dropdown_menu_item]{dropdown_menu_item()}},
\code{\link[=dropdown_menu_label]{dropdown_menu_label()}},
\code{\link[=dropdown_menu_separator]{dropdown_menu_separator()}},
\code{\link[=inc_block_progress]{inc_block_progress()}},
\code{\link[=show_toast]{show_toast()}},
\code{\link[=table_column]{table_column()}},
\code{\link[=update_block_accordion]{update_block_accordion()}},
\code{\link[=update_block_alert_dialog]{update_block_alert_dialog()}},
\code{\link[=update_block_dialog]{update_block_dialog()}},
\code{\link[=update_block_dropdown_menu]{update_block_dropdown_menu()}},
\code{\link[=update_block_popover]{update_block_popover()}},
\code{\link[=update_block_progress]{update_block_progress()}},
\code{\link[=update_block_table]{update_block_table()}},
\code{\link[=update_block_toaster]{update_block_toaster()}}
}
\concept{content}
