% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showcase.R
\name{run_showcase}
\alias{run_showcase}
\title{Run the shinyblocks showcase app}
\usage{
run_showcase(...)
}
\arguments{
\item{...}{Arguments passed to \code{shiny::runApp()}.}
}
\value{
The result of \code{shiny::runApp()}.
}
\description{
Run the shinyblocks showcase app
}
