% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{shinyblocks-package}
\alias{shinyblocks}
\alias{shinyblocks-package}
\title{shinyblocks: Composable modern UI blocks for Shiny}
\description{
\code{shinyblocks} provides composable dashboard primitives for Shiny apps.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://nvelden.github.io/shinyblocks/}
}

}
\author{
\strong{Maintainer}: Niels van der Velden \email{n.s.j.vandervelden@gmail.com}

}
\keyword{internal}
